refactor: created MapGeneratorBase class to use the same logic for maplibre and mapbox
