+++ packages/maplibre-gl-export/src/lib/index.ts
@@ -4,3 +4,4 @@ export { default as PrintableAreaManager } from './printable-area-manager';
 
 export * from './local';
 export * from './interfaces';
+export * from './map-generator-base';
